Fixed database compaction.
To tell the server to compact a database, you send a POST request
to server/database/_compact. Originally, I had been sending the
request without a body (I set the data to send to "", which has a
string length of 0, so no data was even being offered to be sent).
libcurl won't perform POST requests if there is no body; there has
to be *something* being sent. Now, in db_compact(...), I
	set_data(pr, "{}");
to send *something*. From the couchDB docs, the server is supposed
to ignore any data sent as part of the compaction post request,
so I bet I could get away with " " or something silly like that.